Moved the old database backend into a submodule (fallback.py)
