+++ wagtail/search/backends/database/__init__.py
@@ -1,0 +1,8 @@
+from .fallback import DatabaseSearchBackend
+
+
+def SearchBackend(params):
+    """
+    Returns the appropriate search backend for the current 'default' database system
+    """
+    return DatabaseSearchBackend(params)
